feat: Add postExpense function to expenseService
This commit adds the postExpense function to the expenseService module. The function sends a POST request to the "/budget/1/expense/" endpoint to add a new expense. If successful, it returns the response data. If an error occurs, it throws an error message.

Also, update ExpenseForm component

This commit updates the ExpenseForm component by adding a console.log statement to log when an expense is added. This can be useful for debugging purposes.

Update AuthContext

This commit updates the AuthContext module by adding localStorage.setItem("flatshare", JSON.stringify(data)) to store the flatshare data in the local storage when it is fetched successfully. Additionally, it removes the "flatshare" item from the local storage when the user logs out.