refactor: update API endpoints and services (#43)
* refactor: update API endpoints and services

- Update the API endpoint for the todo checklist to "/todo/flat/"
- Add a new API endpoint for the budget at "/budget/"
- Create a new service for fetching flat budgets

The API endpoints and services have been refactored to accommodate the changes in the application. This commit updates the endpoints in the `endpoints.ts` file and adds a new service `budgetService.ts` for fetching flat budgets. The `ExpenseForm` and `ExpenseList` components have also been updated to use the new `ExpenseType` type instead of `Expense`. Additionally, the imports for the checklist services have been corrected to use the proper casing.

* budget: fetch expenses works

* feat: budget get flashare info

* Refactor: Update ExpensePage and AuthContext

- Update ExpensePage to import AuthContext and useContext hook
- Add flatshare property to AuthContext and useContext in ExpensePage
- Display flatshare name in ExpensePage

* fix: bug user in apiFlat a